requirements.txt
asgiref==3.11.0
coverage==7.13.0
dj-database-url==3.0.1
Django==4.2.17
django-cors-headers==4.9.0
django-filter==24.3
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-yasg==1.21.11
factory_boy==3.3.3
Faker==38.2.0
gunicorn==23.0.0
inflection==0.5.1
iniconfig==2.3.0
packaging==25.0
pluggy==1.6.0
psycopg2-binary==2.9.11
Pygments==2.19.2
PyJWT==2.10.1
pytest==9.0.2
pytest-cov==7.0.0
pytest-django==4.11.1
python-dotenv==1.2.1
pytz==2025.2
PyYAML==6.0.3
sqlparse==0.5.4
tzdata==2025.3
uritemplate==4.2.0
whitenoise==6.11.0
